Rename inference methods.
