--- sbi/inference/snl/__init__.py
@@ -1,1 +1,0 @@
-from sbi.inference.snl.snl import SNL
