+++ sbi/inference/snle/__init__.py
@@ -1,0 +1,1 @@
+from sbi.inference.snle.snle_a import SNLE_A  # noqa: F401
