+++ sbi/inference/snle/snle_a.py
@@ -1,0 +1,7 @@
+from sbi.inference.snle.snle_base import LikelihoodEstimator
+
+
+class SNLE_A(LikelihoodEstimator):
+    """SNL=SNLE_A is the only likelihood-based estimator as of now."""
+
+    pass
