+++ sbi/inference/snpe/__init__.py
@@ -1,4 +1,5 @@
-from sbi.inference.snpe.snpe_a import SnpeA
-from sbi.inference.snpe.snpe_b import SnpeB
-from sbi.inference.snpe.snpe_base import SnpeBase
-from sbi.inference.snpe.snpe_c import SnpeC
+from sbi.inference.snpe.snpe_a import SNPE_A  # noqa: F401
+from sbi.inference.snpe.snpe_b import SNPE_B  # noqa: F401
+from sbi.inference.snpe.snpe_base import PosteriorEstimator  # noqa: F401
+from sbi.inference.snpe.snpe_c import SNPE_C  # noqa: F401
+
