+++ sbi/inference/snre/__init__.py
@@ -1,0 +1,6 @@
+from sbi.inference.snre.snre_a import SNRE_A
+from sbi.inference.snre.snre_b import SNRE_B
+
+# Aliases
+AALR = SNRE_A
+SRE = SNRE = SNRE_B
