--- sbi/inference/sre/__init__.py
@@ -1,1 +1,0 @@
-from sbi.inference.sre.sre import SRE
